Move string handling code to new file
